Move bullet cleaner out
Moving functionality out of the well tested bit...

cleaning up

make it green?

Fix tests.

fix test
