+++ packages/frontend/model/extract-capi.ts
@@ -1,4 +1,3 @@
-import { string as curly } from 'curlyquotes';
 import {
     getString,
     getNumber,
@@ -9,11 +8,10 @@ import {
     optional,
 } from './validators';
 import { clean } from './clean';
+import { string as curly } from 'curlyquotes';
+
 import { getSharingUrls } from './sharing-urls';
 import { findPillar } from './find-pillar';
-
-const bigBullets = (s: string) =>
-    s.replace(/•/g, '<span class="bullet">&bull;</span>');
 
 // tslint:disable:prefer-array-literal
 const apply = (input: string, ...fns: Array<(_: string) => string>): string => {
@@ -178,13 +176,12 @@ export const extract = (data: {}): CAPIType => {
         ),
         headline: apply(
             getNonEmptyString(data, 'config.page.headline'),
-            clean,
             curly,
+            clean,
         ),
         standfirst: apply(
             getString(data, 'contentFields.fields.standfirst', ''),
             clean,
-            bigBullets,
         ),
         main: apply(getString(data, 'contentFields.fields.main', ''), clean),
         body: getArray<any>(data, 'contentFields.fields.blocks.body')
